Drop python 3.4 compatibility
closes #41
