+++ setup.cfg
@@ -5,10 +5,9 @@ license_file = LICENSE
 universal = 1
 
 [tool:pytest]
-minversion = 3.3
+minversion = 5
 testpaths = tests
 addopts = --cov
-mock_use_standalone_module = true
 
 [flake8]
 ignore = E126,E128
