+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py{34,35,36,37,38}
+envlist = py{35,36,37,38}
 skip_missing_interpreters = true
 
 [testenv]
